requirements.txt
click
psycopg2-binary
fastapi
uvicorn[standard]
websockets
